unit test and change interface
