+++ process/logsevents/serialize.go
@@ -73,3 +73,22 @@ func serializeDeploy(deployInfo *data.ScDeployInfo) ([]byte, error) {
 
 	return []byte(serializedDataStr), nil
 }
+
+// SerializeTokens will serialize the provided tokens data in a way that Elastic Search expects a bulk request
+func (logsAndEventsProcessor) SerializeTokens(tokens []*data.TokenInfo) ([]*bytes.Buffer, error) {
+	buffSlice := data.NewBufferSlice()
+	for _, tokenData := range tokens {
+		meta := []byte(fmt.Sprintf(`{ "index" : { "_id" : "%s" } }%s`, tokenData.Token, "\n"))
+		serializedData, errMarshal := json.Marshal(tokenData)
+		if errMarshal != nil {
+			return nil, errMarshal
+		}
+
+		err := buffSlice.PutData(meta, serializedData)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	return buffSlice.Buffers(), nil
+}
