+++ process/logsevents/serialize_test.go
@@ -57,3 +57,36 @@ func TestLogsAndEventsProcessor_SerializeSCDeploys(t *testing.T) {
 `
 	require.Equal(t, expectedRes, res[0].String())
 }
+
+func TestSerializeTokens(t *testing.T) {
+	t.Parallel()
+
+	tok1 := &data.TokenInfo{
+		Name:      "TokenName",
+		Ticker:    "TKN",
+		Token:     "TKN-01234",
+		Timestamp: 50000,
+		Issuer:    "erd123",
+		Type:      core.SemiFungibleESDT,
+	}
+	tok2 := &data.TokenInfo{
+		Name:      "Token2",
+		Ticker:    "TKN2",
+		Token:     "TKN2-51234",
+		Issuer:    "erd1231213123",
+		Timestamp: 60000,
+		Type:      core.NonFungibleESDT,
+	}
+	tokens := []*data.TokenInfo{tok1, tok2}
+
+	res, err := (&logsAndEventsProcessor{}).SerializeTokens(tokens)
+	require.Nil(t, err)
+	require.Equal(t, 1, len(res))
+
+	expectedRes := `{ "index" : { "_id" : "TKN-01234" } }
+{"name":"TokenName","ticker":"TKN","token":"TKN-01234","issuer":"erd123","type":"SemiFungibleESDT","timestamp":50000}
+{ "index" : { "_id" : "TKN2-51234" } }
+{"name":"Token2","ticker":"TKN2","token":"TKN2-51234","issuer":"erd1231213123","type":"NonFungibleESDT","timestamp":60000}
+`
+	require.Equal(t, expectedRes, res[0].String())
+}
